[#100] refactor: 상품디테일 구매수량 부분 컴포넌트 분리
