+++ client/components/ProductDetail/ProductInfo/ProductInfo.tsx
@@ -1,17 +1,15 @@
 import React, { ReactElement, useState } from 'react';
 import * as S from './style';
 
-import PlusIcon from '@image/icon/plusIcon.svg';
-import MinusIcon from '@image/icon/minusIcon.svg';
 import {
   DELIVERY_INFO_FEE,
   DELIVERY_INFO_TIMELIMIT,
   TITLE_ORIGIN_AMOUNT,
   TITLE_SALE_AMOUNT,
   TITLE_DELIVERY,
-  TITLE_BUY_AMOUNT,
   TITLE_TOTAL_AMOUNT,
-} from '@client/constants/productDetail/productDetailInfo/productDetailInfo';
+} from '@constants/productDetail/productDetailInfo/productDetailInfo';
+import ProductInfoCount from './ProductInfoCount';
 
 interface Props {
   title: string;
@@ -21,39 +19,6 @@ interface Props {
 
 export default function ProductInfo({ title, originalAmount, amount }: Props): ReactElement {
   const [count, setCount] = useState(1);
-  const [inputValue, setInputValue] = useState<string>(count + '');
-
-  const handleClickCountMinus = () => {
-    if (count <= 1) return;
-    setCount((count) => count - 1);
-    setInputValue((inputValue) => +inputValue - 1 + '');
-  };
-
-  const handleClickCountPlus = () => {
-    if (count + 1 > 100) return;
-    setCount((count) => count + 1);
-    setInputValue((inputValue) => +inputValue + 1 + '');
-  };
-  const handleInputChange = ({ target }: { target: HTMLElement | null }) => {
-    if (!(target instanceof HTMLInputElement)) return;
-    if (+target.value > 100) setInputValue('99');
-    else setInputValue(target.value);
-    setInputValue(target.value);
-  };
-
-  const handleCountSumbit = (e: React.FormEvent) => {
-    e.preventDefault();
-
-    if (+inputValue <= 0) {
-      setCount(1);
-      setInputValue('1');
-    } else if (+inputValue > 100) {
-      setCount(100);
-      setInputValue('100');
-    } else {
-      setCount(+inputValue);
-    }
-  };
 
   return (
     <S.ProductInfo>
@@ -77,16 +42,7 @@ export default function ProductInfo({ title, originalAmount, amount }: Props): R
           </div>
         </div>
         <div className="producto-info__count">
-          <S.InfoTitle>{TITLE_BUY_AMOUNT}</S.InfoTitle>
-          <form onSubmit={handleCountSumbit}>
-            <button type="button" className="count-btn" onClick={handleClickCountMinus}>
-              <MinusIcon />
-            </button>
-            <input type="number" value={inputValue} onChange={handleInputChange} />
-            <button type="button" className="count-btn" onClick={handleClickCountPlus}>
-              <PlusIcon />
-            </button>
-          </form>
+          <ProductInfoCount {...{ count, setCount }} />
         </div>
       </div>
       <div className="product__total-info">
